Added very basic map loader and map file. Restructured assets folder, added folders for images and map files. Cleaned up GenericImage spatial (and subclasses) by loading image in constructor. Updated GameplayState to use map loader. Reduced player image size.
--HG--
rename : assets/goomba.png => assets/images/goomba.png
rename : assets/jar.png => assets/images/jar.png
rename : assets/menu_background.jpg => assets/images/menu_background.jpg
rename : assets/menu_button_exit.png => assets/images/menu_button_exit.png
rename : assets/menu_button_play.png => assets/images/menu_button_play.png
rename : assets/parakoopa.png => assets/images/parakoopa.png
rename : assets/ship.png => assets/images/ship.png
rename : assets/tile.png => assets/images/tile.png

+++ src/net/piemaster/jario/components/Respawn.java
@@ -4,8 +4,8 @@ package net.piemaster.jario.components;
 public class Respawn extends Expires
 {
 	private int initTime;
-	private int respawnX;
-	private int respawnY;
+	private float respawnX;
+	private float respawnY;
 	private boolean active;
 	
 	public Respawn(int initTime)
@@ -13,7 +13,7 @@ public class Respawn extends Expires
 		this(initTime, 0, 0);
 	}
 
-	public Respawn(int initTime, int respawnX, int respawnY)
+	public Respawn(int initTime, float respawnX, float respawnY)
 	{
 		super(initTime);
 		
@@ -22,27 +22,27 @@ public class Respawn extends Expires
 		this.respawnY = respawnY;
 	}
 
-	public int getRespawnX()
+	public float getRespawnX()
 	{
 		return respawnX;
 	}
 
-	public int getRespawnY()
+	public float getRespawnY()
 	{
 		return respawnY;
 	}
 
-	public void setRespawnX(int respawnX)
+	public void setRespawnX(float respawnX)
 	{
 		this.respawnX = respawnX;
 	}
 
-	public void setRespawnY(int respawnY)
+	public void setRespawnY(float respawnY)
 	{
 		this.respawnY = respawnY;
 	}
 	
-	public void setRespawnLocation(int x, int y)
+	public void setRespawnLocation(float x, float y)
 	{
 		this.respawnX = x;
 		this.respawnY = y;
